Revert TLV type change for original_funding_outpoint
The (de)serialisation for `ChannelTransactionParameters` changed
between 0.0.114 and 0.0.116. This type had already been modified by us
when adding the `original_funding_outpoint` field. When first
introduced, we chose a type value of 14.

After the update, we chose to reduce it down to 12 since the type 12
was no longer being used after changes between 0.0.114 and 0.0.116.
Unfortunately, this unnecessarily broke channels created before the
update.

Eventually we will need to settle on a way to handle these possible
conflicts with upstream. Tibo already suggested a few options:

- Use much larger numbers for `rust-dlc` custom fields.
- Move LDK values around as we rebase on top of new changes.
- Add a subdomain for `rust-dlc` custom fields.